Weapon Location/Selection Logic Refactor
The previous implementation of weapon selection, especially the moduleNumber/locationCount parts would be difficult to work with if trying to delete and replace weapons, or later when dealing with battle damage.  They have been refactored with a better system.
The weaponList display elements need to be styled and logic for deleting weapons so they can be replaced needs to be implemented.