Feature and Switch UI and Delete Logic
Logic to delete selected features and switches and revert the changes has been completed.  Styling and display for them has also been completed.  The next step is to create logic to validate all user inputs and create an object holding all ship data that can then be stringified and stored in LocalStorage.