Module Customization Rework 1 - UNFINISHED
I've started reworking the system so that the modules will be stored in a 2d grid and each module will be an object with customization information attached. This approach will allow each module to be customized much more easily and all the modules won't have to be reset is shipSM, shipTL, streamlinedUn, or superScienceChecked change. It should be way better once it's finished.