Module Customization Rework 2 - UNFINISHED
The initial version of the useCallback function to update the module array when one of the dependencies change is finished. It needs testing to be considered complete. getWeaponStats was also updated to be closer to a pure function instead of just a useEffect so it can be reused more easily.