Module Customization Rework 6 - UNFINISHED
The initial version of the habitat customization has been completed. Anytime the SM, TL, or super science changes habitat customizations will be reset. In a future version it would be nice to preserve customizations that are still valid but this will do for this version. More testing would be good for this feature. Next up is weapon customization.