Module Customization Rework 9 - UNFINISHED
The automation level of the habitat modules can now be customized. The initial version of habitat customization is finished. A potential improvement would be to show the workspaces and cost of each habitat as it's being customized rather than having to look at the overall stats. That can wait for an improved version. Some of the work for engine and fuel tank customization has been completed.