Module Customization Rework 13 - UNFINISHED
The bug with the fuel tank checkboxes has been fixed and the styles are functional. The initial version of engine and fuel customization is complete. The UI could use a lot of improvement. The engine accel data, engines, and fuel tanks all look too similar. Some icons or color changes etc might help. Perhaps the UI needs a more radical redesign but the current version will do for now.