Module Customization Rework 17 - UNFINISHED
Some logic errors in module cost calculations were fixed, the div no longer snaps to the top when a checkbox is ticked (it's a bit janky but it's still way better than the previous iteration), and a button to assign all fuel tanks at once was added. A good improvement for that would be to add a state variable that will uncheck all the boxes for that element if the fuel tanks don't all have the same contents but the current version where it matches the first tank will do for now. Plus some initial work for weapon customization.